Move table function operator to package
Same as other non-trivial operators where helper classes exist.
