+++ core/trino-main/src/main/java/io/trino/operator/function/TableFunctionPartition.java
@@ -11,8 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.operator;
+package io.trino.operator.function;
 
+import io.trino.operator.WorkProcessor;
 import io.trino.spi.Page;
 
 public interface TableFunctionPartition
